Implemented untested autonomous code refactor
