+++ Team6220_PowerPlay/src/main/java/org/firstinspires/ftc/team6220_PowerPlay/competition/RedRight.java
@@ -2,17 +2,20 @@ package org.firstinspires.ftc.team6220_PowerPlay.competition;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
+import org.firstinspires.ftc.team6220_PowerPlay.AutoFramework;
 import org.firstinspires.ftc.team6220_PowerPlay.BaseAutonomous;
 import org.firstinspires.ftc.team6220_PowerPlay.Constants;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraRotation;
 
 @Autonomous(name = "RedRight")
-public class RedRight extends BaseAutonomous {
+public class RedRight extends AutoFramework {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        // initialize motors, servos, imu, cameras, etc.
+        runAuto(1);
+
+        /*// initialize motors, servos, imu, cameras, etc.
         initialize();
 
         // detect april tag in initialize
@@ -73,6 +76,8 @@ public class RedRight extends BaseAutonomous {
             case 2:
                 driveAutonomous(-90, 33);
                 break;
+
+         */
         }
-    }
+
 }
